refactor : move screen content to nutrition screen
